+++ src/main/java/gift/controller/api/WishController.java
@@ -1,10 +1,10 @@
-package gift.controller;
+package gift.controller.api;
 
 import gift.annotation.CurrentMember;
-import gift.dto.WishRequestDto;
+import gift.dto.wish.WishRequestDto;
 import gift.dto.wish.WishResponseDto;
 import gift.entity.Member;
-import gift.service.WishService;
+import gift.service.wish.WishService;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
