+++ src/main/java/gift/dto/member/TokenResponseDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.member;
 
 public class TokenResponseDto {
