+++ src/main/java/gift/dto/wish/WishResponseDto.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.wish;
 
 public class WishResponseDto {
