+++ src/main/java/gift/exception/GlobalExceptionHandler.java
@@ -2,6 +2,7 @@ package gift.exception;
 
 import gift.exception.member.MemberException;
 import gift.exception.product.ProductException;
+import gift.exception.productOption.ProductOptionException;
 import gift.exception.wishList.WishException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -57,6 +58,15 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(WishException.class)
     public ResponseEntity<Map<String, String>>
     handlerValidationError(WishException e) {
+        Map<String, String> errors = new HashMap<>();
+        errors.put("message", e.getMessage());
+
+        return ResponseEntity.status(e.getHttpStatus()).body(errors);
+    }
+
+    @ExceptionHandler(ProductOptionException.class)
+    public ResponseEntity<Map<String, String>>
+    handlerValidationError(ProductOptionException e) {
         Map<String, String> errors = new HashMap<>();
         errors.put("message", e.getMessage());
